Reset timer upon launching batch-based run
This isn't a huge deal, but it helps in the case where a batch run (i.e.
enough nodes are in the buffer to prompt a run) occurs just before a
tick otherwise would. We now restart the timer under such circumstances,
so the next timer-based check will occur <duration> after the batch run,
rather than immediately.